pyproject.toml
[project]
name = "weather-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "fastapi[standard]>=0.133.1",
    "httpx>=0.28.1",
    "opentelemetry-exporter-otlp-proto-http>=1.28.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "opentelemetry-instrumentation-httpx>=0.49b0",
    "opentelemetry-instrumentation-redis>=0.49b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.49b0",
    "opentelemetry-sdk>=1.28.0",
    "prometheus-client>=0.24.1",
    "pydantic-settings>=2.13.1",
    "python-json-logger>=4.0.0",
    "redis>=7.2.1",
    "sqlalchemy[asyncio]>=2.0.47",
    "sqlmodel>=0.0.37",
]

[dependency-groups]
dev = [
    "anyio>=4.12.1",
    "pytest>=9.0.2",
    "pytest-benchmark>=5.2.3",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.4",
]
